backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib
pydantic-settings
email-validator
bcrypt
openai
